+++ src/vs/workbench/api/browser/mainThreadDocumentsAndEditors.ts
@@ -32,7 +32,7 @@ import { diffSets, diffMaps } from '../../../base/common/collections.js';
 import { IPaneCompositePartService } from '../../services/panecomposite/browser/panecomposite.js';
 import { ViewContainerLocation } from '../../common/views.js';
 import { IConfigurationService } from '../../../platform/configuration/common/configuration.js';
-import { IDirtyDiffModelService } from '../../contrib/scm/browser/dirtyDiff.js';
+import { IDirtyDiffModelService } from '../../contrib/scm/browser/dirtyDiffModel.js';
 
 
 class TextEditorSnapshot {
