+++ src/vs/workbench/api/test/browser/mainThreadDocumentsAndEditors.test.ts
@@ -36,7 +36,7 @@ import { LanguageService } from '../../../../editor/common/services/languageServ
 import { ILanguageConfigurationService } from '../../../../editor/common/languages/languageConfigurationRegistry.js';
 import { TestLanguageConfigurationService } from '../../../../editor/test/common/modes/testLanguageConfigurationService.js';
 import { IUndoRedoService } from '../../../../platform/undoRedo/common/undoRedo.js';
-import { IDirtyDiffModelService } from '../../../contrib/scm/browser/dirtyDiff.js';
+import { IDirtyDiffModelService } from '../../../contrib/scm/browser/dirtyDiffModel.js';
 import { ITextEditorDiffInformation } from '../../../../platform/editor/common/editor.js';
 
 suite('MainThreadDocumentsAndEditors', () => {
