+++ src/vs/workbench/contrib/scm/browser/dirtyDiffModel.ts
